requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
python-multipart==0.0.12
Pillow==10.4.0
pytesseract==0.3.13
pdf2image==1.17.0
python-docx==1.1.2
opencv-python-headless==4.9.0.80
numpy==1.26.4
httpx==0.27.0
openai==1.12.0
